BUGFIX: FINALLY, AWESOME!
This commit doesn't add anything new; just wanted to emphasize that the
previous commit really DID fix the problem with Tc calculations. It was
the order of setting the current system temperature vs. loading the
material states (with associated posthooks) that messed up the Tc
calculations, and resulted in flat segments in the otherwise monotonic
curve, i.e. which effectively lowered the resolution of the output.

Hopefully, there are no more new hidden problems with the new code...